+++ quasar.conf.js
@@ -8,6 +8,7 @@ module.exports = function (ctx) {
     // --> boot files are part of "main.js"
     // https://quasar.dev/quasar-cli/cli-documentation/boot-files
     boot: [
+      'settings-storage',
       'i18n',
       'flespi-io',
       'icomoon',